src/requirements.txt
Pillow==11.0.0
numba==0.60.0
